feat: replace hand-written db-types with Drizzle-inferred types [E06-W3a]
Move 5 table schema definitions (entities, transactions, budgets,
inventory, wishlist) from pops-api into @pops/db-types as the single
source of truth. Replace Zod-based hand-written row types with
InferSelectModel/InferInsertModel from drizzle-orm.

Key changes:
- db-types now depends on drizzle-orm instead of zod
- New @pops/db-types/schema export with table objects
- pops-api schema files re-export from @pops/db-types
- Row types are now camelCase (matching Drizzle column names)
- Removed 6 toRow() snake_case mapping functions from services
- Updated all to*() mapper functions for camelCase properties
- Added EntityInsert/TransactionInsert/etc. insert types
- Net -318 lines (111 added, 429 removed)

All 498 tests pass, typecheck clean across pops-api and pops-shell.
