+++ apps/pops-api/src/modules/finance/wishlist/service.ts
@@ -8,22 +8,6 @@ import { wishList } from "../../../db/schema/wishlist.js";
 import { NotFoundError } from "../../../shared/errors.js";
 import type { WishListRow, CreateWishListItemInput, UpdateWishListItemInput } from "./types.js";
 
-/** Map a Drizzle select result back to the snake_case WishListRow expected by the router. */
-type DrizzleWishList = typeof wishList.$inferSelect;
-function toRow(r: DrizzleWishList): WishListRow {
-  return {
-    id: r.id,
-    notion_id: r.notionId,
-    item: r.item,
-    target_amount: r.targetAmount,
-    saved: r.saved,
-    priority: r.priority,
-    url: r.url,
-    notes: r.notes,
-    last_edited_time: r.lastEditedTime,
-  };
-}
-
 /** Count + rows for a paginated list. */
 export interface WishListListResult {
   rows: WishListRow[];
@@ -56,8 +40,7 @@ export function listWishListItems(
     .orderBy(asc(wishList.item))
     .limit(limit)
     .offset(offset)
-    .all()
-    .map(toRow);
+    .all();
 
   const [{ total }] = db
     .select({ total: count() })
@@ -78,7 +61,7 @@ export function getWishListItem(id: string): WishListRow {
     .get();
 
   if (!row) throw new NotFoundError("Wish list item", id);
-  return toRow(row);
+  return row;
 }
 
 /**
