+++ apps/pops-api/src/modules/inventory/items/types.ts
@@ -26,28 +26,28 @@ export interface InventoryItem {
   lastEditedTime: string;
 }
 
-/** Map a SQLite row to the API response shape. */
+/** Map a database row to the API response shape. */
 export function toInventoryItem(row: InventoryRow): InventoryItem {
   return {
     id: row.id,
-    itemName: row.item_name,
+    itemName: row.itemName,
     brand: row.brand,
     model: row.model,
-    itemId: row.item_id,
+    itemId: row.itemId,
     room: row.room,
     location: row.location,
     type: row.type,
     condition: row.condition,
-    inUse: row.in_use === 1,
+    inUse: row.inUse === 1,
     deductible: row.deductible === 1,
-    purchaseDate: row.purchase_date,
-    warrantyExpires: row.warranty_expires,
-    replacementValue: row.replacement_value,
-    resaleValue: row.resale_value,
-    purchaseTransactionId: row.purchase_transaction_id,
-    purchasedFromId: row.purchased_from_id,
-    purchasedFromName: row.purchased_from_name,
-    lastEditedTime: row.last_edited_time,
+    purchaseDate: row.purchaseDate,
+    warrantyExpires: row.warrantyExpires,
+    replacementValue: row.replacementValue,
+    resaleValue: row.resaleValue,
+    purchaseTransactionId: row.purchaseTransactionId,
+    purchasedFromId: row.purchasedFromId,
+    purchasedFromName: row.purchasedFromName,
+    lastEditedTime: row.lastEditedTime,
   };
 }
