--- packages/db-types/src/entities.ts
@@ -1,31 +1,0 @@
-/**
- * SQLite schema for entities table (snake_case columns).
- * Used by pops-api for entity types.
- */
-import { z } from "zod/v4";
-
-/** Supported entity types — extensible via new values, not schema changes. */
-export const ENTITY_TYPES = [
-  "company",
-  "person",
-  "place",
-  "brand",
-  "organisation",
-] as const;
-
-export type EntityType = (typeof ENTITY_TYPES)[number];
-
-export const EntityRowSchema = z.object({
-  id: z.string(),
-  notion_id: z.string().nullable(),
-  name: z.string(),
-  type: z.string(),
-  abn: z.string().nullable(),
-  aliases: z.string().nullable(),
-  default_transaction_type: z.string().nullable(),
-  default_tags: z.string().nullable(),
-  notes: z.string().nullable(),
-  last_edited_time: z.string(),
-});
-
-export type EntityRow = z.infer<typeof EntityRowSchema>;
