+++ packages/db-types/src/schema/index.ts
@@ -1,0 +1,5 @@
+export { entities } from "./entities.js";
+export { transactions } from "./transactions.js";
+export { budgets } from "./budgets.js";
+export { homeInventory } from "./inventory.js";
+export { wishList } from "./wishlist.js";
